% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmedio.R
\name{pmedio}
\alias{pmedio}
\title{Integración con la regla del punto medio:}
\usage{
pmedio(f, a, b)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Esta función calcula la integral de una función dada f entre dos puntos a y b
mediante la regla del punto medio.
}
