% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangulo.R
\name{rectangulo}
\alias{rectangulo}
\title{Integración con la regla del rectángulo:}
\usage{
rectangulo(f, a, b)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Esta función calcula la integral de una función dada f entre dos puntos a y b
mediante la regla del rectángulo.
}
